# [OPT] Opt resize performance and  do image hash before resize