--- a/docker/Dockerfile.byted
+++ b/docker/Dockerfile.byted
@@ -52,6 +52,9 @@ RUN http_proxy='' pip install absl-py==${absl_version} --extra-index-url https:/
 # group_gemm 安装
 RUN pip install --verbose git+https://github.com/bytedance-iaas/grouped_gemm@dev
 
+# opencv install
+RUN http_proxy='' pip install opencv-python --extra-index-url https://bytedpypi.byted.org/simple
+
 # examples 覆盖
 RUN cd /tmp && git clone https://github.com/bytedance-iaas/sglang.git -b ${branch} && \
     rm -rf /sgl-workspace/sglang/examples && \
@@ -84,6 +87,10 @@ RUN USE_TRITON_DISTRIBUTED_AOT=0 pip3 install -e python --verbose --no-build-iso
 RUN curl -fsSL https://mirrors.byted.org/extra-tools/ubuntu/GPG-KEY-system | gpg --dearmor -o /etc/apt/trusted.gpg.d/volc-extra-tools.gpg && \
     echo "deb [arch=$(dpkg --print-architecture) signed-by=/etc/apt/trusted.gpg.d/volc-extra-tools.gpg] http://mirrors.byted.org/extra-tools/ubuntu jammy main" > /etc/apt/sources.list.d/volctools.list && \
     apt update && apt install -y onion-ai-data
+
+# libgl install
+RUN apt update
+RUN apt install -y libgl1
 
 # entrypoint
 COPY docker/entrypoint.sh /entrypoint.sh