<!-- Thank you for your contribution! Please follow these guidelines to enhance your pull request. If anything is unclear, submit your PR and reach out to maintainers for assistance. Join our Slack community at https://slack.sglang.ai to discuss further. -->

## Motivation
1. opt resize peformance
2. opt hash logic 

## Modifications

<!-- Detail the changes made in this pull request. -->

## Accuracy Tests

<!-- If this pull request affects model outputs (e.g., changes to the kernel or model forward code), provide accuracy test results. -->

## Benchmarking and Profiling

before:
[2025-09-02 14:14:59] INFO:     127.0.0.1:57892 - "POST /v1/chat/completions HTTP/1.1" 200 OK
The image appears to be the cover of a children's book titled "小さいおばけ" (Koai Obae), which translates to "Little
[2025-09-02 14:14:59] INFO:     127.0.0.1:57892 - "POST /v1/chat/completions HTTP/1.1" 200 OK
[2025-09-02 14:14:59] INFO:     127.0.0.1:57892 - "POST /v1/chat/completions HTTP/1.1" 200 OK
cost time 23.578643798828125 ms
[2025-09-02 14:14:59] INFO:     127.0.0.1:57892 - "POST /v1/chat/completions HTTP/1.1" 200 OK
cost time 40.351152420043945 ms
[2025-09-02 14:15:00] INFO:     127.0.0.1:57892 - "POST /v1/chat/completions HTTP/1.1" 200 OK
cost time 512.4974250793457 ms
[2025-09-02 14:15:01] INFO:     127.0.0.1:57892 - "POST /v1/chat/completions HTTP/1.1" 200 OK
cost time 1092.7579402923584 ms
[2025-09-02 14:15:02] INFO:     127.0.0.1:57892 - "POST /v1/chat/completions HTTP/1.1" 200 OK
cost time 1185.459852218628 ms
Action_Summary: 左键

after:

The image appears to be a cover of a book or comic titled "おはげ" (Ohage), which translates to "Gap Teeth" in English.
[2025-09-02 14:21:31] INFO:     127.0.0.1:57908 - "POST /v1/chat/completions HTTP/1.1" 200 OK
[2025-09-02 14:21:31] INFO:     127.0.0.1:57908 - "POST /v1/chat/completions HTTP/1.1" 200 OK
cost time 23.4220027923584 ms
[2025-09-02 14:21:31] INFO:     127.0.0.1:57908 - "POST /v1/chat/completions HTTP/1.1" 200 OK
cost time 28.12027931213379 ms
[2025-09-02 14:21:31] INFO:     127.0.0.1:57908 - "POST /v1/chat/completions HTTP/1.1" 200 OK
cost time 221.16446495056152 ms
[2025-09-02 14:21:32] INFO:     127.0.0.1:57908 - "POST /v1/chat/completions HTTP/1.1" 200 OK
cost time 535.1457595825195 ms
[2025-09-02 14:21:33] INFO:     127.0.0.1:57908 - "POST /v1/chat/completions HTTP/1.1" 200 OK
cost time 638.4117603302002 ms
Action_Summary: 左键

## Checklist

- [ ] Format your code according to the [Format code with pre-commit](https://docs.sglang.ai/developer_guide/contribution_guide.html#format-code-with-pre-commit).
- [ ] Add unit tests according to the [Run and add unit tests](https://docs.sglang.ai/developer_guide/contribution_guide.html#run-and-add-unit-tests).
- [ ] Update documentation according to [Write documentations](https://docs.sglang.ai/developer_guide/contribution_guide.html#write-documentations).
- [ ] Provide accuracy and speed benchmark results according to [Test the accuracy](https://docs.sglang.ai/developer_guide/contribution_guide.html#test-the-accuracy) and [Benchmark the speed](https://docs.sglang.ai/developer_guide/contribution_guide.html#benchmark-the-speed).